core/config.py
"""Configuration management."""
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

class Config:
    """Application configuration."""
    
    # Paths
    BASE_DIR = Path(__file__).parent.parent
    LOGS_DIR = BASE_DIR / "logs"
    
    # Model settings
    MODEL_NAME = os.getenv("MODEL_NAME", "qwen2.5-coder: 7b")
    OLLAMA_HOST = os.getenv("OLLAMA_HOST", "http://localhost:11434")
    
    # Server settings
    HOST = os.getenv("HOST", "0.0.0.0")
    PORT = int(os.getenv("PORT", "8000"))
    
    # Logging
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
    
    @classmethod
    def ensure_dirs(cls):
        """Create required directories."""
        cls. LOGS_DIR.mkdir(exist_ok=True)

Config.ensure_dirs()