requirements.txt
chainlit==1.1.0
ollama==0.1.7
pydantic==2.10.1
python-dotenv==1.0.0
colorlog==6.8.0